pyproject.toml
[project]
name = "logistic-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Aleksandr Danilenko", email = "rorik302@gmail.com"},
]
dependencies = [
    "podman-compose>=1.0.6",
    "litestar[pydantic,sqlalchemy,standard]>=2.6.0",
    "asyncpg>=0.29.0",
    "argon2-cffi>=23.1.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"

[tool.pdm.dev-dependencies]
lint = [
    "pre-commit>=3.6.0",
    "ruff>=0.2.1",
    "black>=24.1.1",
]

[tool.ruff]
select = ["E", "W", "F", "C", "B", "I"]
line-length = 120

[tool.black]
line-length = 120

[tool.pdm.scripts]
cli = "python src/cli/main.py"
